% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{read_qpgraph}
\alias{read_qpgraph}
\title{Import a qpGraph file into an admixturegraph object}
\usage{
read_qpgraph(text)
}
\arguments{
\item{text}{Text containing the graph description.}
}
\value{
An admixturegraph object
}
\description{
Import a qpGraph file into an admixturegraph object
}
